Refactor code for Alpaca-style prompts
