+++ blacksmith/datasets/torch/custom/custom_dataset_utils.py
@@ -1,15 +1,19 @@
 # SPDX-FileCopyrightText: (c) 2026 Tenstorrent AI ULC
 #
 # SPDX-License-Identifier: Apache-2.0
-from blacksmith.datasets.torch.custom.templates.alpaca import (
-    ALPACA_INPUT_BLOCK_TEMPLATE,
-    ALPACA_PROMPT_TEMPLATE,
-)
+from enum import Enum
+
+from blacksmith.datasets.torch.custom.templates.alpaca import build_alpaca_prompt
 
 
-def build_alpaca_prompt(instruction: str, input_text: str = "") -> str:
-    input_section = ALPACA_INPUT_BLOCK_TEMPLATE.substitute(input=input_text) if input_text.strip() else ""
-    return ALPACA_PROMPT_TEMPLATE.substitute(
-        instruction=instruction,
-        input_section=input_section,
-    )
+class AvailableFormats(Enum):
+    ALPACA = "alpaca"
+
+
+def build_prompt(format: str, instruction: str, input_text: str = ""):
+    format = format.strip().lower()
+    if format == AvailableFormats.ALPACA.value:
+        return build_alpaca_prompt(instruction, input_text)
+    else:
+        available_formats = [f.value for f in AvailableFormats]
+        raise ValueError(f"Unsupported format: {format}. Available formats: {available_formats}")
